+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `players` property on `GameSate` for listing all players in the game
 - `find_team` and `find_player` methods on `GameState`
 - Alias `Player#name` as `Player#username`
+- `Team.Color#by_name`
 
 ## [0.1.1] - 2018-11-06
 ### Fixed